contact entity acl check on api system related #844

+++ src/Ojs/ApiBundle/Controller/Admin/ContactRestController.php
@@ -5,6 +5,7 @@ namespace Ojs\ApiBundle\Controller\Admin;
 use FOS\RestBundle\Controller\Annotations\Get;
 use FOS\RestBundle\Controller\FOSRestController;
 use Nelmio\ApiDocBundle\Annotation\ApiDoc;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\HttpKernel\Exception\HttpException;
 use FOS\RestBundle\Controller\Annotations\View;
 use Ojs\AdminBundle\Form\Type\ContactType;
@@ -40,6 +41,9 @@ class ContactRestController extends FOSRestController
      */
     public function getContactsAction(Request $request, ParamFetcherInterface $paramFetcher)
     {
+        if (!$this->isGranted('VIEW', new JournalContact())) {
+            throw new AccessDeniedHttpException;
+        }
         $offset = $paramFetcher->get('offset');
         $offset = null === $offset ? 0 : $offset;
         $limit = $paramFetcher->get('limit');
@@ -68,6 +72,9 @@ class ContactRestController extends FOSRestController
     public function getContactAction($id)
     {
         $entity = $this->getOr404($id);
+        if (!$this->isGranted('VIEW', $entity)) {
+            throw new AccessDeniedHttpException;
+        }
         return $entity;
     }
 
@@ -85,6 +92,9 @@ class ContactRestController extends FOSRestController
      */
     public function newContactAction()
     {
+        if (!$this->isGranted('CREATE', new JournalContact())) {
+            throw new AccessDeniedHttpException;
+        }
         return $this->createForm(new ContactType(), null, ['csrf_protection' => false]);
     }
 
@@ -106,6 +116,9 @@ class ContactRestController extends FOSRestController
      */
     public function postContactAction(Request $request)
     {
+        if (!$this->isGranted('CREATE', new JournalContact())) {
+            throw new AccessDeniedHttpException;
+        }
         try {
             $newEntity = $this->container->get('ojs_api.contact.handler')->post(
                 $request->request->all()
@@ -141,6 +154,9 @@ class ContactRestController extends FOSRestController
      */
     public function putContactAction(Request $request, $id)
     {
+        if (!$this->isGranted('CREATE', new JournalContact())) {
+            throw new AccessDeniedHttpException;
+        }
         try {
             if (!($entity = $this->container->get('ojs_api.contact.handler')->get($id))) {
                 $statusCode = Codes::HTTP_CREATED;
@@ -189,6 +205,9 @@ class ContactRestController extends FOSRestController
                 $this->getOr404($id),
                 $request->request->all()
             );
+            if (!$this->isGranted('EDIT', $entity)) {
+                throw new AccessDeniedHttpException;
+            }
             $routeOptions = array(
                 'id' => $entity->getId(),
                 '_format' => $request->get('_format')
@@ -224,6 +243,9 @@ class ContactRestController extends FOSRestController
     public function deleteContactAction($id)
     {
         $entity = $this->getOr404($id);
+        if (!$this->isGranted('DELETE', $entity)) {
+            throw new AccessDeniedHttpException;
+        }
         $this->container->get('ojs_api.contact.handler')->delete($entity);
         return $this->view(null, Codes::HTTP_NO_CONTENT, []);
     }
